Alter the anchors when merging files, adding a prefix

+++ t/compile-merged.t
@@ -15,7 +15,7 @@ binmode $builder->todo_output,    ":utf8";
 binmode STDOUT, ':encoding(utf-8)';
 binmode STDERR, ':encoding(utf-8)';
 
-my $testnum = 132;
+my $testnum = 140;
 
 my $xelatex = $ENV{TEST_WITH_LATEX};
 if ($xelatex) {
@@ -71,6 +71,14 @@ like $outtex, qr/\\setmainlanguage\{french\}/, "Found language selection";
 like $outtex, qr/\\setotherlanguages\{.*russian.*\}/, "Found russian lang";
 like $outtex, qr/\\setotherlanguages\{.*english.*\}/, "Found english lang";
 like $outtex, qr/\\russianfont/, "Found russian font";
+
+foreach my $pnum (1..4) {
+    like $outtex, qr{hyperref\{\}\{piece00000\Q$pnum\Eamuse\}},
+      "Found hyperref with $pnum piece";
+    like $outtex, qr{hyperdef\{piece00000\Q$pnum\Eamuse\}},
+      "Found hyperdef with $pnum piece";
+}
+
 
 if ($xelatex) {
     ok(-f "$base.pdf", "$base.pdf created");
@@ -145,10 +153,12 @@ like $epub_html, qr{My <em>new</em> shiny test}, "Found the title";
 like $epub_html, qr{Another <em>one</em>}, "Found the author";
 like $epub_html, qr{<em>Today</em>!}, "Found the date";
 
-foreach my $ext (qw/aux log pdf tex toc status epub/) {
-    my $remove = "$base.$ext";
-    if (-f $remove) {
-        unlink $remove or warn $!;
+unless ($ENV{NO_CLEANUP}) {
+    foreach my $ext (qw/aux log pdf tex toc status epub/) {
+        my $remove = "$base.$ext";
+        if (-f $remove) {
+            unlink $remove or warn $!;
+        }
     }
 }
 
@@ -167,6 +177,7 @@ sub _get_epub_xhtml {
         my $html_piece = read_file(File::Spec->catfile($tmpdir->dirname, $piece));
         # neutralize internal linking for testing purposes
         $html_piece =~ s/(href=")piece[0-9]+\.xhtml(#.*?")/$1$2/g;
+        $html_piece =~ s/(text-amuse-label-)piece[0-9]+-/$1/g;
         push @html, "<!-- $piece -->\n", $html_piece;
     }
     return join('', @html);
